requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
requests>=2.31.0
python-telegram-bot>=20.0
solana>=0.30.0
solders>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
web3>=6.0.0
eth-account>=0.11.0
mnemonic>=0.21
bip32>=3.0
pynacl>=1.5.0
construct>=2.10.68
